requirements.txt
torch
torchvision
pycocotools
numpy
matplotlib
Pillow
scikit-learn
